// client/src/api/models/postApiIdentityRegisterBody.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Readdit
 * OpenAPI spec version: v1
 */

export type PostApiIdentityRegisterBody = {
   FirstName: string;
   LastName: string;
   Username: string;
   Email: string;
   EmailConfirmationUrl: string;
   Password: string;
   Country: string;
   Gender: string;
   ProfilePicture?: Blob;
};